+++ src/extensions/update_manifest_handlers/steps_handler/src/steps_handler.cpp
@@ -78,9 +78,9 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
     ADUC_Result result = { .ResultCode = ADUC_Result_Failure, .ExtendedResultCode = 0 };
     ADUC_WorkflowHandle childHandle = nullptr;
 
-    auto stepCount = static_cast<unsigned int>(workflow_get_instructions_steps_count(handle));
+    auto stepCount = static_cast<size_t>(workflow_get_instructions_steps_count(handle));
     char* workFolder = workflow_get_workfolder(handle);
-    unsigned int childWorkflowCount = workflow_get_children_count(handle);
+    size_t childWorkflowCount = workflow_get_children_count(handle);
     ADUC_FileEntity* entity = nullptr;
     int workflowLevel = workflow_get_level(handle);
 
@@ -95,8 +95,8 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
             workflow_free(child);
         }
 
-        Log_Debug("Creating workflow for %d step(s). Parent's level: %d", stepCount, workflowLevel);
-        for (unsigned int i = 0; i < stepCount; i++)
+        Log_Debug("Creating workflow for %lu step(s). Parent's level: %d", stepCount, workflowLevel);
+        for (size_t i = 0; i < stepCount; i++)
         {
             STRING_HANDLE childId = nullptr;
             childHandle = nullptr;
@@ -150,7 +150,7 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
                 catch (...)
                 {
                     Log_Error(
-                        "Exception occurred while downloading a detached Update Manifest file for level#%d step#%d (file id:%s).",
+                        "Exception occurred while downloading a detached Update Manifest file for level#%d step#%lu (file id:%s).",
                         workflowLevel,
                         i,
                         entity.FileId);
@@ -164,7 +164,7 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
                 if (IsAducResultCodeFailure(result.ResultCode))
                 {
                     Log_Error(
-                        "An error occurred while downloading manifest file for step#%d (erc:%d)",
+                        "An error occurred while downloading manifest file for step#%lu (erc:%d)",
                         i,
                         result.ExtendedResultCode);
                     goto done;
@@ -188,7 +188,7 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
                         JSON_Value* compsValue = nullptr;
                         if (compatibilityString.get() == nullptr)
                         {
-                            Log_Error("Cannot get compatibility info for components-update #%d", i);
+                            Log_Error("Cannot get compatibility info for components-update #%lu", i);
                             result = { .ResultCode = ADUC_Result_Failure,
                                        .ExtendedResultCode =
                                            ADUC_ERC_STEPS_HANDLER_GET_REF_STEP_COMPATIBILITY_FAILED };
@@ -200,7 +200,7 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
 
                         if (IsAducResultCodeFailure(result.ResultCode))
                         {
-                            Log_Error("Cannot select components for components-update #%d", i);
+                            Log_Error("Cannot select components for components-update #%lu", i);
                             goto done;
                         }
 
@@ -227,7 +227,7 @@ ADUC_Result PrepareStepsWorkflowDataObject(ADUC_WorkflowHandle handle)
             }
             else
             {
-                childId = STRING_construct_sprintf("%d", i);
+                childId = STRING_construct_sprintf("%lu", i);
                 workflow_set_id(childHandle, STRING_c_str(childId));
                 STRING_delete(childId);
                 childId = nullptr;
@@ -550,12 +550,12 @@ static ADUC_Result StepsHandler_Download(const tagADUC_WorkflowData* workflowDat
         // For each step (child workflow), invoke backup, install and apply actions.
         // if install or apply fails, invoke restore action.
         //
-        for (int i = 0; i < stepsCount; i++)
+        for (size_t i = 0; i < stepsCount; i++)
         {
             if (IsStepsHandlerExtraDebugLogsEnabled())
             {
                 Log_Debug(
-                    "Perform download action of child step #%d on component #%d.\n#### Component ####\n%s\n###################\n",
+                    "Perform download action of child step #%lu on component #%d.\n#### Component ####\n%s\n###################\n",
                     i,
                     iCom,
                     serializedComponentString);
@@ -567,7 +567,7 @@ static ADUC_Result StepsHandler_Download(const tagADUC_WorkflowData* workflowDat
             stepHandle = workflow_get_child(handle, i);
             if (stepHandle == nullptr)
             {
-                const char* errorFmt = "Cannot process step #%d due to missing (child) workflow data.";
+                const char* errorFmt = "Cannot process step #%lu due to missing (child) workflow data.";
                 Log_Error(errorFmt, i);
                 result.ExtendedResultCode = ADUC_ERC_STEPS_HANDLER_DOWNLOAD_FAILURE_MISSING_CHILD_WORKFLOW;
                 workflow_set_result_details(handle, errorFmt, i);
@@ -582,7 +582,7 @@ static ADUC_Result StepsHandler_Download(const tagADUC_WorkflowData* workflowDat
                 {
                     result.ResultCode = ADUC_Result_Failure;
                     result.ExtendedResultCode = ADUC_ERC_STEPS_HANDLER_SET_SELECTED_COMPONENTS_FAILURE;
-                    workflow_set_result_details(handle, "Cannot select target component(s) for step #%d", i);
+                    workflow_set_result_details(handle, "Cannot select target component(s) for step #%lu", i);
                     goto done;
                 }
             }
@@ -592,13 +592,13 @@ static ADUC_Result StepsHandler_Download(const tagADUC_WorkflowData* workflowDat
                 ? workflow_peek_update_manifest_step_handler(handle, i)
                 : DEFAULT_REF_STEP_HANDLER;
 
-            Log_Info("Loading handler for step #%d (handler: '%s')", i, stepUpdateType);
+            Log_Info("Loading handler for step #%lu (handler: '%s')", i, stepUpdateType);
 
             result = ExtensionManager::LoadUpdateContentHandlerExtension(stepUpdateType, &contentHandler);
 
             if (IsAducResultCodeFailure(result.ResultCode))
             {
-                const char* errorFmt = "Cannot load a handler for step #%d (handler :%s)";
+                const char* errorFmt = "Cannot load a handler for step #%lu (handler :%s)";
                 Log_Error(errorFmt, i, stepUpdateType);
                 workflow_set_result(stepHandle, result);
                 workflow_set_result_details(handle, errorFmt, i, stepUpdateType == nullptr ? "NULL" : stepUpdateType);
@@ -833,7 +833,7 @@ static ADUC_Result StepsHandler_Install(const tagADUC_WorkflowData* workflowData
         // For each step (child workflow), invoke backup, install and apply actions.
         // if install or apply fails, invoke restore action.
         //
-        for (int i = 0; i < stepsCount; i++)
+        for (size_t i = 0; i < stepsCount; i++)
         {
             if (IsStepsHandlerExtraDebugLogsEnabled())
             {
@@ -850,7 +850,7 @@ static ADUC_Result StepsHandler_Install(const tagADUC_WorkflowData* workflowData
             stepHandle = workflow_get_child(handle, i);
             if (stepHandle == nullptr)
             {
-                const char* errorFmt = "Cannot process step #%d due to missing (child) workflow data.";
+                const char* errorFmt = "Cannot process step #%lu due to missing (child) workflow data.";
                 Log_Error(errorFmt, i);
                 result.ExtendedResultCode = ADUC_ERC_STEPS_HANDLER_INSTALL_FAILURE_MISSING_CHILD_WORKFLOW;
                 workflow_set_result_details(handle, errorFmt, i);
@@ -875,13 +875,13 @@ static ADUC_Result StepsHandler_Install(const tagADUC_WorkflowData* workflowData
                 ? workflow_peek_update_manifest_step_handler(handle, i)
                 : DEFAULT_REF_STEP_HANDLER;
 
-            Log_Info("Loading handler for child step #%d (handler: '%s')", i, stepUpdateType);
+            Log_Info("Loading handler for child step #%lu (handler: '%s')", i, stepUpdateType);
 
             result = ExtensionManager::LoadUpdateContentHandlerExtension(stepUpdateType, &contentHandler);
 
             if (IsAducResultCodeFailure(result.ResultCode))
             {
-                const char* errorFmt = "Cannot load a handler for step #%d (handler :%s)";
+                const char* errorFmt = "Cannot load a handler for step #%lu (handler :%s)";
                 Log_Error(errorFmt, i, stepUpdateType);
                 workflow_set_result(stepHandle, result);
                 workflow_set_result_details(handle, errorFmt, i, stepUpdateType == nullptr ? "NULL" : stepUpdateType);
@@ -1311,7 +1311,7 @@ static ADUC_Result StepsHandler_IsInstalled(const tagADUC_WorkflowData* workflow
         serializedComponentString = CreateComponentSerializedString(selectedComponentsArray, iCom);
 
         // For each step (child workflow), invoke IsInstalled().
-        for (int i = 0; i < stepsCount; i++)
+        for (size_t i = 0; i < stepsCount; i++)
         {
             if (IsStepsHandlerExtraDebugLogsEnabled())
             {
@@ -1328,7 +1328,7 @@ static ADUC_Result StepsHandler_IsInstalled(const tagADUC_WorkflowData* workflow
             stepHandle = workflow_get_child(handle, i);
             if (stepHandle == nullptr)
             {
-                const char* errorFmt = "Cannot process child step #%d due to missing (child) workflow data.";
+                const char* errorFmt = "Cannot process child step #%lu due to missing (child) workflow data.";
                 Log_Error(errorFmt, i);
                 result.ExtendedResultCode = ADUC_ERC_STEPS_HANDLER_ISINSTALLED_FAILURE_MISSING_CHILD_WORKFLOW;
                 workflow_set_result_details(handle, errorFmt, i);
@@ -1343,7 +1343,7 @@ static ADUC_Result StepsHandler_IsInstalled(const tagADUC_WorkflowData* workflow
                 {
                     result.ResultCode = ADUC_Result_Failure;
                     result.ExtendedResultCode = ADUC_ERC_STEPS_HANDLER_SET_SELECTED_COMPONENTS_FAILURE;
-                    workflow_set_result_details(handle, "Cannot set target component(s) for child step #%d", i);
+                    workflow_set_result_details(handle, "Cannot set target component(s) for child step #%lu", i);
                     goto done;
                 }
             }
@@ -1353,13 +1353,13 @@ static ADUC_Result StepsHandler_IsInstalled(const tagADUC_WorkflowData* workflow
                 ? workflow_peek_update_manifest_step_handler(handle, i)
                 : DEFAULT_REF_STEP_HANDLER;
 
-            Log_Debug("Loading handler for child step #%d (handler: '%s')", i, stepUpdateType);
+            Log_Debug("Loading handler for child step #%lu (handler: '%s')", i, stepUpdateType);
 
             result = ExtensionManager::LoadUpdateContentHandlerExtension(stepUpdateType, &contentHandler);
 
             if (IsAducResultCodeFailure(result.ResultCode))
             {
-                const char* errorFmt = "Cannot load a handler for child step #%d (handler :%s)";
+                const char* errorFmt = "Cannot load a handler for child step #%lu (handler :%s)";
                 Log_Error(errorFmt, i, stepUpdateType);
                 workflow_set_result_details(handle, errorFmt, i, stepUpdateType == nullptr ? "NULL" : stepUpdateType);
                 goto done;
@@ -1397,7 +1397,7 @@ static ADUC_Result StepsHandler_IsInstalled(const tagADUC_WorkflowData* workflow
                 || result.ResultCode == ADUC_Result_IsInstalled_NotInstalled)
             {
                 Log_Info(
-                    "Workflow lvl %d, step #%d, child step #%d, component #%d is not installed.",
+                    "Workflow lvl %d, step #%d, child step #%lu, component #%d is not installed.",
                     workflowLevel,
                     workflowStep,
                     i,
